back-end/ClothesTryOn/requirements.txt
configobj==5.0.6
flask==1.0.2
h5py==2.8.0
ipython==5.7.0
keras==2.1.6
matplotlib==2.2.2
numpy==1.16.0
opencv-python==3.4.2.16
pandas==0.23.0
Pillow==5.1.0
pyzmq==17.0.0
scipy==1.1.0
six==1.11.0
tensorflow==1.14.0
djangorestframework==3.10.3
